Move audio components to their own package
