+++ collect_app/src/main/java/org/odk/collect/android/audio/AudioPlayerViewModel.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.viewmodel;
+package org.odk.collect.android.audio;
 
 import android.content.Context;
 import android.media.MediaPlayer;
