+++ collect_app/src/test/java/org/odk/collect/android/audio/AudioPlayerViewModelTest.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.viewmodel;
+package org.odk.collect.android.audio;
 
 import android.media.MediaPlayer;
 import android.net.Uri;
